pyproject.toml
[project]
name = "dtlabs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "boto3>=1.37.8",
    "oci>=2.147.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "python-dotenv>=1.0.1",
]
